Gameover, startscreen and playscreen
all three of these screens are displaying everything as intended, but the highscore number is still 223 lol.

+++ Framework-2025/SDL_Template/StartScreen.h
@@ -101,7 +101,7 @@ public:
 
 	void ChangeSelectedMode(int change);
 
-	void SetHighScore(int score);
+	void SetHighScore(float score);
 
 	void ReloadHighScore();
